feat(#191): 옵션 버튼 공통 컴포넌트로 분리
